Finished YAML based configurationa and parsing Fixed issue with GETCommits endpoint Fixed other DI related issues
Partial Config implementation

[deploy: development]

+++ src/main/java/com/ftinc/lolserv/data/model/LolCommit.java
@@ -158,7 +158,7 @@ public class LolCommit implements ModelMap<LolCommit> {
         commitHash   = (String) map.get("hash");
         optionalKey  = (String) map.get("optional_key");
         imageUrl     = (String) map.get("image_url");
-        timestamp    = (long) map.get("timestamp");
+        timestamp    = (int) map.get("timestamp");
         return this;
     }
